requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
anthropic==0.28.0
python-dotenv==1.0.1
pydantic==2.7.1
